Remove the evex512 target attribute
The evex512 attribute was introduced in GCC 14
https://gcc.gnu.org/pipermail/gcc-patches/2023-September/631081.html
when Intel was considering the introduction of AVX10.1-256, that is,
a 256-bit subset of AVX512. The __attribute__((target("evex512")))
would "opt in" to 512-bit operations.

Later, Intel canceled the plans to differentiate between AVX10.1-256
and AVX10.1-512, and the AVX10 feature set will always be 512-bit.

The EVEX instruction prefix will remain in use, but the evex512
attribute was removed in GCC 16 along with the traces of avx10.1-xxx:
https://gcc.gnu.org/pipermail/gcc-patches/2025-May/683609.html